[hack] Initial work on a full-fidelity hack parser
Summary:
The existing Hack parser was designed primarily for use by the type checker, and does not have "full fidelity" with the source code; for a variety of reasons we wish to have a parser which definitively assigns every character in a file to a position in the parse tree. Text which is not part of parse tree nodes, like comments and whitespace, is assigned "trivia" nodes.

This is just a first cut; nowhere near complete. We'll be fleshing out the lexer and parser to support the full language, adding test cases, and adding features like syntactic rewriters.

Differential Revision: D3273489

fbshipit-source-id: 9b93b850ae5f697addcb37ec68b0dc53b51e5d80