fix json incremental bug
Summary:
We were neglecting to check whether an unsatisfied import/require might later resolve to a JSON file.
This meant that we failed to recheck possible dependents when a JSON file appeared incrementally.

In the other direction, we were incorrectly resolving JSON module paths to SourceFile filenames.
This meant that we failed to recheck dependents of a JSON file when it disappeared incrementally.

This diff fixes both and moves the if-statement that checks for a JSON file extension when building a Loc.filename
into to a helper function - the logic was replicated in a couple of places already, and this fix would have added
a couple more.

New test in incremental_json

Reviewed By: mroch

Differential Revision: D3351524

fbshipit-source-id: dc63ce75b3b1885864001fe04ab9f969ee052bb6