Make workers share stderr with the master
Summary: Before D2796468 worker processes shared stdout and stderr with the master process. Sharing at least stderr is very useful for making it easy to run hh_server in a console window and watch for the occasional status message or alarm showing up in the console. This does not preclude workers from logging copious amount so data to individual log files if they so choose. However, without this diff workers do not get to choose to write things to the stderr of the master.

Reviewed By: gabelevi

Differential Revision: D3349980

fbshipit-source-id: 2d250337a5a107df9902c0793711b1f35771f1bb