[hack] Full fidelity "trivia" nodes; whitespace, newlines, comments
Summary:
This diff includes the simplest data structures used in the full-fidelity parse tree:

* A module to represent the abstract notion of source text
* A module that helps map line/column offsets to absolute offsets and vice-versa
* Modules to represeint "trivia" -- whitespace, newlines, comments.

Source text is for now just a thin wrapper around a string and an offset map. Eventually we'll have more information in here, like file name, and so on.

The offset map is a linear search; this is not very good. Working up a better data structure here is possibly a good task for one of our summer interns.

As we will see in later diffs, my aim here is to provide various different implementations of abstractions such as "trivia", "token" and "syntax tree node".

Each will have different performance characteristics and capabilities. For this initial set of diffs, there are three main capabilities:

* "minimal" nodes are designed to be as small in memory as possible, with low burden to the memory allocator, and capable of being serialized into a small format.
* "editable" nodes are designed for scenarios in which a syntax tree is being manipulated programmatically to construct new trees.
* "positioned" nodes are heavyweight, and know their exact position in a source text.

Minimal trivia nodes consist of a kind and a width; that's it. They don't know their own text or their position. As we'll see in later diffs, we can compute both by descending the parse tree.

Editable trivia nodes consist of a kind and their text. They don't know their positions; indeed, as the tree is being edited, the positions of every node might change.

Positioned nodes know their absolute offset, their width, and what source text they came from.

Reviewed By: dabek

Differential Revision: D3325553

fbshipit-source-id: ea26204e4f5435c5bf78cdeea2e310d805bf4a4a