Update React docs
Summary:
Removes text that stateless functional component support is coming soon. Adds
examples of stateless functional components, including using destructuring with
default values as default props.

Updates HOC example to be an actualy HOC, i.e., a function. Goes into a bit more
detail about the HOC and fleshes out the example.

Adds jump-to links at the top of the page, which will hopefully make it clearer
what we support.

Removes usage of experimental class properties. Instead, just declares the
properties and assigns in constructor. Seems fine.

Removes the animated gif, which we thought was cool but kind of distracting and
disjointed with the rest of the docs.

Closes https://github.com/facebook/flow/issues/1081
Closes https://github.com/facebook/flow/issues/1082
Closes https://github.com/facebook/flow/issues/1523
Closes https://github.com/facebook/flow/issues/1601

Reviewed By: mroch

Differential Revision: D3369711

fbshipit-source-id: 962ba9584d84b42188211ef7054427978a394dc3