Fix sentinel prop refinement when prop is open
Summary:
Before this diff we would give up rather quickly when looking to do a sentinel
property refinement. Specifically, if the l-value has a prop, but it is open, it
might eventually resolve to be a suitable refinement.

This diff creates a new use type which waits for an open property type to become
concrete before either pruning the branch or expanding the result type.

Fixes https://github.com/facebook/flow/issues/1827

Reviewed By: avikchaudhuri

Differential Revision: D3370429

fbshipit-source-id: 644a2752e059a39c73ead078e61a28a75d9c3219