[hack] Remove recursive types from Hack
Summary:
This is the first stage of removing recursive types from Hack.

In a new module typing_unify_recursive, the "Env.add env x ty" function to add a substitution to the environment is
wrapped with an "occurs check", testing whether x appears inside ty. If it is, instead of making the assignment,
an error message is produced.

However, for some very specific cases we successfully unify types without introducing recursion at all. For example,
when unifying ?v with v for some type variable v, instead
of creating a recursive type, we assign v := ?u for some fresh variable u.
This is possible because ? is idempotent: ??t = ?t for any t, and in the
absence of recursive types this is the only valid solution to the problem.

(This also works for repeated use of ? e.g. ??v unified with v).

See https://fb.facebook.com/groups/721152551318888/permalink/777622745671868/ for more details.

Reviewed By: ericlippert

Differential Revision: D3075705

fbshipit-source-id: e00dc26ea752f74fca828cc49e71844776a0593e