[parser][easy] don't lex an extra token when calling Lookahead.junk
Summary:
`Lookahead.junk` calls `lex_until ... 1` but `lex_until` uses `<=` so `lex_until ... 1` actually lexes 2 tokens.

this meant that every time we advance the parser and mean to throw out the first looked-ahead token, we were also looking ahead at the next token. next time we peeked we'd just use it, so no big deal, but this could be an issue if there isn't an extra token (EOF).

Reviewed By: bhosmer

Differential Revision: D3383447

fbshipit-source-id: 22543c20c5d25cbddac9f339988f8f22ba28845d