[parser] consolidate references to lexbuf
Summary:
minor (but kinda scary) cleanup to use the `lexbuf` in `env.lex_env` instead of additionally storing it in `env.lb`.

when an `env` is created, a `lex_env` is also created and both `env.lb` and `env.lex_env.lex_lb` are set to the same value. AFAICT, these two stay in sync (verifying this is the scary part).

consolidating them elimiates the chance that they get out of whack.

Reviewed By: gabelevi

Differential Revision: D3280999

fbshipit-source-id: 51cd2f70024251cce88bcd5883bd503b93a1ac18