[parser] clean up regexp/template parsing, remove Eat.advance
Summary:
created modes for regexp and template lexing, which cleaned up a bunch of stuff.

i wasn't sure whether it mattered that the token stream JSON output changed, but if it does matter, I think we should hack it so that it stringifies them as "normal" (legit since they're both only valid in normal mode).

Reviewed By: gabelevi

Differential Revision: D3383543

fbshipit-source-id: da37dfa11d1599667e26d9854bdb803293fc42b4