[hack] Support recursive constraints between generic parameters
Summary:
Currently, the naming phase rejects constraints on *classes* such as
```
class C<T1 as T2, T2 as T1>
```
There are two bugs here though:
* Similar cycles on *method* generic parameters are not rejected
* In such a method, it's possible to induce looping in the type checker.

This diff removes the naming phase check completely, and instead spots looping during subtype checking. For example, consider checking `T1 <: D` under the constraints above. The subtype algorithm will recurse with `T2 <: D` and then with `T1 <: D`. At this point, it will recognise that it has seen T1 before, and raise an error. Note that the size of a cycle is bounded by the number of generic parameters, so we could just count instead of keeping a visited-set. But this makes error messages silly, as they might repeat parameters e.g. if there were three more parameters T3, T4, and T5 above, not in the cycle, then we would go round the cycle three times before hitting the max count.

There is no need to reject cycles in constraints, and in future, the ability to require two types to be equal (which is what you get from T1 as T2, T2 as T1) will be useful, for example, under safe instanceof, or through extension of constraints on type parameters at method scope.

Reviewed By: dabek

Differential Revision: D3358852

fbshipit-source-id: de2077e028767f780b5422e4937ae8977729497f