Implement server-side coverage command
Summary:
Running `flow coverage` on the file below, in the Nuclide codebase, occupies the Flow server for two minutes on my machine:

`Nuclide/pkg/nuclide-reprint-js/lib/printers/complex/printLiteral.js`

This is because coverage is implemented using the `dump-types` server command, which forces all of the types to be stringified and sent across the process boundary. The file above, for whatever reason, has types that are ridiculously long when stringified. To solve this, I've added a new `coverage` command to the server, which simply returns a boolean depending on whether the region is covered or not. With this change, running `flow coverage` on the file above is basically instantaneous, and the output is identical to the output without this change.

I will follow up with another PR to add detailed coverage information to the JSON output, so Nuclide can switch from dump-types to coverage for its coverage display. Then we can get away from this pathological behavior.

Of course, this PR doesn't fix `dump-types
Closes https://github.com/facebook/flow/pull/1903

Differential Revision: D3396462

Pulled By: mroch

fbshipit-source-id: 6a10d17701ee4ec5c6114127ba417d08cfdaa1ed