[hack] Get source code line extents given an offset
Summary: I am working on some parse tree error reporting code; for the unit test I am probably going to display the line of code with the error.  This requires being able to extract a line from the source code given its offset. These are a couple of helper methods to do so.

Reviewed By: alexchow

Differential Revision: D3405670

fbshipit-source-id: 90199d2f5605eaf0c4926d051540ec395761ac26