fix tracking of const types after abnormal control flow
Summary:
Our current approach to control handling involves clearing
the flow-sensitive ('specific') types of bound variables
to EmptyT after abnormal control flow, e.g. early return.
Pending a change to this approach, consts must participate
in this operation. This is at odds with other aspects of
const behavior, including their immunity to havoc.

The latter was causing a bug where a function defined
after an early return was seeing const bindings with
type EmptyT, rather than their conservative (flow-
insensitive) types.

This diff extends the temporary threading of the needle
by recognizing that a const's specific type has been
set to EmptyT - which will only happen in a clearing
operation - and in this case performing a havoc as it
would on a non-const variable. Regression test added.

Reviewed By: mroch

Differential Revision: D3408175

fbshipit-source-id: 87def94c8250eecff9d8848761f97a8080d56b8b