[website] fix blog post headers
Summary:
we already show the title at the top of a post, so including it as a header is redundant. this removes the header and then bumps up all the other header sizes.

the "Recent Posts" list in the nav doesn't always have space for long titles, so added a `short-title` front-matter property. this differs from the "slug" because the slug defines the URL.

Reviewed By: samwgoldman

Differential Revision: D3425580

fbshipit-source-id: a8a6d2a21e1848e0229b8ec40c6eb6dae7147e0e