[Flow] Add .gitattributes to enforce eof=lf
Summary: For the most part, it doesn't matter if our source code uses LF or
CRLF as line terminators. However some of our tests report character offsets,
and these are affected. We could enforce LF for only our various tests, but we
might as well do it repo wide.

This is a GitHub only commit

Reviewed By: bhosmer

Differential Revision: D3432025