Address the float_of_string issue
Summary:
float_of_string is broken for Windows. It is fixed in 4.03.0, but
ocpwin currently only goes up to 4.02.1. So we need some solution.

It probably would be feasible to figure out the licenses in order to pull in
the c implementation of float_of_string from the ocaml project itself. This is
what https://github.com/facebook/flow/pull/1186 tried to do. However
and easier solution is to just hack together a rough float_of_string in OCaml.

As I understand it, the problem with float_of_string is only for hexidecimal
representations of floats. So my solution only needs to parse those. It tries
the pervasives implementation first, and only falls back to my manual
implementation on Windows when an exception is thrown

For reference I used the ocaml implementation to see what hex representations
ocaml supports:
https://github.com/ocaml/ocaml/blob/0323df74a7ccad9804a6da1e1b3d9aa8c3c3c4e8/byterun/floats.c#L189

Reviewed By: samwgoldman

Differential Revision: D3434339

fbshipit-source-id: 42ff2b004617dd58effe7af884e041d092333baa