[website] enable line numbers in fenced code blocks
Summary:
extends the fenced code block (triple-backticks) syntax to support a `+line_numbers` modifier which turns on line numbers just for that code block.

this lets us use fenced code blocks instead of the `{% highlight %}` macro.

the tradeoff is that it overrides a bunch of Kramdown internals, which might make it slightly harder to upgrade kramdown one day. I think it's nice.

Reviewed By: samwgoldman

Differential Revision: D3448208

fbshipit-source-id: 162edb99a7fb486f4a30a4d6c9b9489a7f8086e1