[website] fix huge font sizes on tablets
Summary:
the `<meta name='viewport' content='width=480'>` we were sending causes mobile browsers to make the page 480px wide and then zoom the viewport to fit that. so in portrait, it was zoomed out more than we expected, and it was zoomed in when in landscape.

instead, using its normal width lets the `media` queries work properly. I added a `media` query for tablets and phones in landscape mode.

Fixes https://github.com/facebook/flow/issues/1809

Reviewed By: samwgoldman

Differential Revision: D3449649

fbshipit-source-id: 3493d972fc533886ef1e5254909dbbaef88067ae