[website] automatically render CLI output
Summary:
the example CLI output occasionally falls out of sync. also, I want to show CLI output below every example with errors, and in the future allow the viewer to toggle them on and off (and, perhaps, have them always on on mobile, where the hover isn't easy to use).

this diff reconstructs the CLI output from JSON in the same way that Try Flow does, and injects it below each example. if you don't want that on a specific example, add the `$NoCLIOutput` pragma to the top of that example.

Reviewed By: gabelevi

Differential Revision: D3454446

fbshipit-source-id: ca3f4a5e6f611d1f248f967a9bd1ffb1c5f1a113