Throw when dep or hash tables are full
Summary:
Instead of looping forever, let's blow up when the tables fill up.

We already had a counter for the hash table, it just wasn't being decremented
when we remove from the hash table. The dep table needed it's own counter.

I also changed these from ints to uint64s, since with pow=26 we're getting
close to the line. Might as well future proof a little.

Reviewed By: andrewjkennedy

Differential Revision: D3452498

fbshipit-source-id: 08c5ac58041a8848ced5ea82f91ae1e115ab792e