Inference support for ES "declare module"s
Summary:
This build the inference component for the syntax added in D3321152.

Right now we do this by storing a tvar for the `ModuleT` in the context as we run through the body of the `declare module`.

I would like to come back and also store the outer `ModuleT` tvar in the context in this way as well (and just coordinate that tvar with the modulemap only in the `infer_ast` function), but doing so in this diff gets pretty messy -- so I've saved it for a follow up.

Reviewed By: gabelevi

Differential Revision: D3451288

fbshipit-source-id: c6d1a058f3daa7406729d94a095795b1912fe549