[website] hide auto-generated cli output behind a link on desktop
Summary:
the cli output is fairly redundant with the inline errors on desktop, where you can easily hover over them.

this diff hides the errors on devices that can hover via `media (hover)`.

the ruby changes are just shuffling around when we render stuff, so that we can wrap the code and cli output in a div instead of being separate sections.

Reviewed By: samwgoldman

Differential Revision: D3465943

fbshipit-source-id: 7ef013fb6191658fa4f702ab3a647df09307214e