shortcut quick error case on fun -> obj
Summary:
There's a known issue with the fun -> obj rule that avoids the mechanism for
union type checking.

This diff doesn't fix the actual issue, but implements a solution that makes the
issue not likely to be hit for 99% usecases of fun -> obj.

Fixes https://github.com/facebook/flow/issues/1948

Reviewed By: samwgoldman

Differential Revision: D3454313

fbshipit-source-id: ea66cadb6a36bfd54065bcd541859c378d15e263