Support sharedmemory-log-level option in flow
Summary:
The Hack server now has support for logging statistics about the
bytes serialized to and from the shared memory heap. The infrastructure
is shared by Flow. This diff introduces an option sharedmemory-log-level
which when set to non-zero value enables logging for Flow too.

See D3455947 for more details of what is logged.

Reviewed By: gabelevi

Differential Revision: D3475693

fbshipit-source-id: e8306a1d902442b61f394f1894d7a1627711ac6b