Split up maybe and optional types during concretization
Summary:
The existing code to concretize unions makes it possible to pass a union into an
overload where each branch of the union matches one of the overloads.

This commit also concretizes maybe and optional types, splitting into branches
as unions do. This makes it possible to pass maybe types into overloads that
handle null, void, and the type. Ditto optionals into overloads that handle void
and the type.

Reviewed By: avikchaudhuri

Differential Revision: D3448276

fbshipit-source-id: 35950916b2d44c169ea47006ac4f1a4963bf764f