Reposition conditional prop lookups, not inside OrT/AndT
Summary:
GetPropT results in a reposition, but LookupT does not. The `is_cond` flag to
expression switches between GetPropT and LookupT to perform non-strict lookups
in conditional expressions. This diff adds the repositioning logic from GetPropT
to these non-strict lookups, but not to LookupT in general.

OrT and AndT reposition their left and right hand sides, which worked around the
lack of LookupT repositioning, but now that we do reposition on non-strict
lookups, we don't need this.

Reviewed By: bhosmer

Differential Revision: D3482751

fbshipit-source-id: 837ee23a510f22984963a3e7e40969b4c7edb1e6