Pass traces through recursive flow_opt calls
Summary:
I was trying to track down some weird behavior that hit a typeapp. The
verbose-indent output was all over the place and --traces omitted some
interesting information.

Flow_js exports a number of functions that necessarily call `flow_opt` because
non-recursive callers don't have a trace to pass in. However, if we're calling
recursively and we do have a trace, passing it through adds some valuable
history information to the trace.

Reviewed By: mroch

Differential Revision: D3510548

fbshipit-source-id: ae111e3641291d60a7c211450ef5118c1911b10d