fix refinements involving template literals
Summary:
we only refined typeof and string literal checks when the RHS was a string, not a template literal.

this works now:

```
if (typeof foo === `string`) {...}
if (foo === `bar`) {...}
```

note that it only matches templates with a single quasi, which also implies that there are no expressions.

Fixes https://github.com/facebook/flow/issues/2030

Reviewed By: gabelevi

Differential Revision: D3517676

fbshipit-source-id: 5c1dd6337a61f546c7ef66b1f54056e5351847ce