[appveyor] cache ocpwin installer in S3
Summary:
downloads the ocpwin installer from S3 instead of from ocamlpro since S3 is way faster (like, 5 min faster).

to be on the safe side of licensing, the S3 bucket requires authentication so that we're not redistributing the binary.

the URL is stored as a secure environment variable in appveyor, which means we can only do this for commits to the main repo, not PRs or forks.

Reviewed By: gabelevi

Differential Revision: D3529823

fbshipit-source-id: 785ff776229f3c2b855a1c49dd276bf0d808bdfe