use flowconfig sharedmem settings in convert command
Summary:
`SharedMem.init_default` and `SharedMem.default_config` use Hack's sharedmem settings instead of ours. Let's stop using them.

Added `shm_global_size` and `shm_heap_size` to `FlowConfig.Opts.t` because that's the source of truth for all the other shm settings and this rounds it out, but they are not exposed in `.flowconfig` (we could, didn't see the need).

Reviewed By: gabelevi

Differential Revision: D3532498

fbshipit-source-id: c81269ae56d1eb8669a8435cdab2e386993b79ba