[website] link to Stack Overflow on the Support tab
Summary:
I've been really impressed with the answers to Stack Overflow questions about flow, and GitHub makes it pretty hard to tell the difference between questions and bugs, or to find answers to previously-asked questions; it's also difficult to know where to look when a question involves multiple projects, like how to integrate Flow and Babel.

So I'd like to encourage questions to go to Stack Overflow. GitHub is still a valid place, but I think people will get better answers on SO.

Reviewed By: bhosmer

Differential Revision: D3530590

fbshipit-source-id: 5c24c96ab94660dff999548daa00877beb17d2a9