rename files like Env_js, Reason_js to Env, Reason
Summary:
we've stopped depending on/vendoring in most of the Hack files that conflicted with ours, forcing us to use crappy `_js` extensions.

it's nice to get rid of the stuff like `module Reason = Reason_js`.

this renames a bunch, but there are more. this diff just was getting big.

Reviewed By: gabelevi

Differential Revision: D3539088

fbshipit-source-id: 085e5c5743070556f510c7f2ed537039750fd27e