shortcut some unify cases
Summary:
Report of Flow slowness at:
https://github.com/jlongster/flow-bug-repro

This was caused (among several interacting things) by repeated unification of
similar things that involved an intersection, that hit the naive_unify case
(emitting bidirectional flow constraints). Because the cache is cleared when
trying branches of an intersection, the number of constraints exploded.

With this diff, the checking time goes from 15s to 1s.

Reviewed By: jeffmo

Differential Revision: D3554586

fbshipit-source-id: 2b51955bcc0fedd7bb961b4420e31424d3f50f9c