fix traces for unify
Summary:
For any call to __flow(l,u) with trace, we should have the nice property that
the end points of trace are l and u.

With unification, a slight generalization of this property should still hold:
the end points should be equal modulo equality constraints. (With more explicit
steps that record equality constraint propagation, we won't need this
generalization, but that's for later.)

Unfortunately there was a gross bug in trace management for rec_unify calls that
didn't do the corresponding rec_trace operation on traces. This caused several
levels of trace information to be lost, and the end points of a trace
accompanying rec_unify t1 t2 ended up being arbitrary ancestors.

This diff fixes, along with some refactoring of rec_unify to closely match
rec_flow (adding wrappers around a core __unify function for trace management
and recursion check).

Reviewed By: bhosmer

Differential Revision: D3561325

fbshipit-source-id: 608c7aed309040fa28c7568714e3b2dd624c36cc