covariant class type
Summary:
Previously, Class<S> was a subtype of Class<T> only when S = T. However as it
turns out, ES6 classes actually ensure that not only the instance types, but
also the class types, are related by prototype inheritance. In other words, we
should relax the previous rule to only need that S be a subtype of T.

Unfortunately this invalidates certain uses of Class<.> that relied on the
exactness of previous rule. Ideally this would be solved by providing a type
operator that enforces exactness, but as an alternative this diff lifts an
existing operator to the surface syntax that can replace Class<.> in (some of?)
those uses.

Reviewed By: jeffmo

Differential Revision: D3598558

fbshipit-source-id: e7c03c15dd8cb6b77c91a5751c7c167b67c2c5f6