Store type params in env
Summary:
A while back we added a parameter to nearly every function to support nested
type parameters. This diff stores type parameters in the env structure instead
of threading the params in a map.

Class and function signatures are a bit trickier, since they are processed in
two stages. These modules do hold on to a map of type params, which store the
type params until their `toplevels` functions are called.

To support this, the type annotation module functions continue to accept a map
so annotations can be converted to types for use in classes and functions.

Note that classes don't actually install their own type parameters, but rather
rely on Func_sig.toplevels to do it. I added a test to ensure this works in the
expected way.

Type params are stored as TypeT, same as type aliases. When these are
instantiated with EmptyT, we create an EmptyT -> ReposUpperT constraint, which
was over-eagerly pruned in ground_subtypes, so I added an exception.

Reviewed By: bhosmer

Differential Revision: D3516014

fbshipit-source-id: 264c930248679e43fde25aa5b94e21d240ab271d