Allow object destructuring by computed key
Summary:
It's possible to destructure an object by a computed key. Naturally the key
needs to be rebound to a valid identifier, but we should still allow it.

Destructuring a computed property is analagous to computed property access. The
implementation here uses the same typing rules for `{ [expr]: p } = o` as we
would use for `var p = o[expr]` -- that is, statically literal strings are
treated like props, elem type is checked against a dictionary, if present, unknown
strings and numbers are allowed, and everything else is an error.

Reviewed By: gabelevi

Differential Revision: D3513355

fbshipit-source-id: 8b5d75b81ca8eaebfdf618f46ab5008ce59c1247