Add rate limiting to connection code
Summary:
I have had reports of `flow status` and other commands spin-waiting for the server to initialize or recheck. This should not happen and I have not been able to repro. The `connect_once` function should block trying to connect to the server for a second. However, there's no harm in being defensive here.
Closes https://github.com/facebook/flow/pull/2115

Reviewed By: samwgoldman

Differential Revision: D3602038

Pulled By: gabelevi

fbshipit-source-id: 5142e59c11fce3ad9d84eb8a563521f42b79520d