Handle exceptional behavior in ./tool tests
Summary:
Thus far, `./tool test` expected a test to either pass or fail. If
there's a bug in the `./test.js` file that causes a JS error to be thrown, the
test runner would let it bubble up. If you're running more than one test, then
it sucks that one test can stop the others from running. It was kind of ok for
just `./tool test`, but for `./tool test --watch`, it sucks if you stop
watching just because one test was malformed.

So now a test can pass, fail or error. If a test doesn't even parse, then
`./tool test` will fail immediately, but `./tool test --watch` will just
complain and run the other tests

Reviewed By: samwgoldman

Differential Revision: D3606425

fbshipit-source-id: 35919fe8dd98d1e4c48e7bb2a32a9e4f50bcd9c0