readme.md should be more focused on installing & usage.
Summary:
I've decided to finally add Flow to our dev stack on the project I am currently working on. I had a bit of trouble getting started though, as I was a bit surprised that the first thing I saw in the readme is how to build Flow instead of how to actually install it and get up & running.

For me personally it was quiet stressful as I wasn't sure if I need to do the building or I can just install using brew (I'm pretty sure there's a diagnose for that).

I am not very good at writing stuff so just take this as a poke & point to the better (didn't want to use "right") direction.

Thanks,

Foxhoundn
Closes https://github.com/facebook/flow/pull/2084

Reviewed By: samwgoldman

Differential Revision: D3561958

Pulled By: gabelevi

fbshipit-source-id: 4b6cfa72530512bb14b3928b13f8625840aa5d42