Reposition union-like types in resolved tvars
Summary:
D3569370 caused a regression in repositioning in a test. The changes in the diff
were not the cause, but rather exposed an existing bug in the implementation.

Some types have special repositioning logic, but repositioning a resolved tvar
bypassed that logic and instead just did a naive reposition. Consider a tvar
that was resolved to a union type --- the members of the union would not be
repositioned!

This diff ensures that all repositioning logic is hit for all repositioned
types. The only complex part of this change is dealing with loops, which is
(hopefully) well explained by the inline comments.

Reviewed By: avikchaudhuri

Differential Revision: D3618894

fbshipit-source-id: 776e24e87a48968f97a7f79445894004fb15800d