exact types initial support
Summary:
Introduces a basic setup for handling exact types. Only
exact object types are supported in this diff, but it
should be possible to incrementally add support for the
exact forms of other kinds.

$Exact<T> represents the exact interpretation of T (just
members of T, no subtypes). For a given T,

	Exact<T> can flow to Exact<T>
	Exact<T> can flow to T
	T may *not* flow to Exact<T>

Constants have exact types, so also:

	(constant of type T) can flow to Exact<T>

For examples, see `newtests/exact`.

TODO:
1. Add non-experimental syntax
2. Add support for other kinds as needed

Reviewed By: avikchaudhuri

Differential Revision: D3607502

fbshipit-source-id: 999b88cafb5ba5bb2799f257849b542daccd0136