[hack] Introduce a thread local cache
Summary: This is the first diff towards restructing how we store class type information in the shared heap. The caching we have for the shared heap is also useful in the general case for computation we would like to cache for a given process. This refactors the code a bit to allow for utilizing the caching code, without requiring storing anything in shared memory. In theory this should probably live in a separate folder/file but since the primary user is still the shared heap I think its fine to leave it as is for.

Reviewed By: andrewjkennedy

Differential Revision: D3538528

fbshipit-source-id: 3d2e964f823b57a6b127669126c81b85931f36d6